refactor: update project structure and remove unused files; enhance sidebar and layout components
