+++ components/side-bar.tsx
@@ -1,4 +1,4 @@
-'use server';
+'use client';
 
 import { motion } from 'framer-motion';
 import Image from 'next/image';
