--- pages/api/hello.ts
@@ -1,8 +1,0 @@
-import { NextApiRequest, NextApiResponse } from 'next';
-
-export default function handler(
-  reques: NextApiRequest,
-  response: NextApiResponse,
-) {
-  response.status(200).json({ text: 'Hello' });
-}
